feat: 결제 진행 메세지 출력 기능 추가

+++ src/main/java/domain/Orders.java
@@ -22,7 +22,7 @@ public class Orders {
     }
 
     public int countChickenType() {
-        return (int) orders.keySet()
+        return orders.keySet()
                 .stream()
                 .filter(Menu::isChickenType)
                 .mapToInt(menu -> orders.get(menu))
